# docs/api-docs.yaml
openapi: 3.0.0
info:
  title: Kairo Sync API
  version: 1.0.0
  description: Event management platform with authentication, profiles, events, and templates
servers:
  - url: http://localhost:5000/api
    description: Development server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    EventRequest:
      type: object
      required: [name, startDate, startTime]
      properties:
        name:
          type: string
          maxLength: 200
        location:
          type: string
          maxLength: 300
        lat:
          type: number
          minimum: -90
          maximum: 90
        lng:
          type: number
          minimum: -180
          maximum: 180
        visibility:
          type: string
          enum: [public, private]
          default: public
        capacity:
          type: integer
          minimum: 1
        additionalNotes:
          type: string
          maxLength: 1000
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        startTime:
          type: string
          format: time
        endTime:
          type: string
          format: time
        inviteUserIds:
          type: array
          items:
            type: string
        coverPic:
          type: string
          format: binary
...
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: object
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
        error:
          type: string
  responses:
    Success:
      description: Operation successful
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
    Created:
      description: Resource created successfully
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
    BadRequest:
      description: Invalid request data
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
paths:
...
  /events:
    get:
      tags: [Events]
      summary: Get events with filters
      security:
        - BearerAuth: []
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [all, public, private, hosted, invited, upcoming]
            default: all
        - name: filter
          in: query
          schema:
            type: string
            enum: [recommended, recent, distant, nearest, farthest]
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
        - name: search
          in: query
          schema:
            type: string
        - name: userLat
          in: query
          schema:
            type: number
            minimum: -90
            maximum: 90
        - name: userLng
          in: query
          schema:
            type: number
            minimum: -180
            maximum: 180
        - name: sortBy
          in: query
          schema:
            type: string
        - name: sortOrder
          in: query
          schema:
            type: string
            enum: [ASC, DESC]
            default: DESC
      responses:
        '200':
          $ref: '#/components/responses/Success'
        '401':
          $ref: '#/components/responses/Unauthorized'
    
    post:
      tags: [Events]
      summary: Create new event
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/EventRequest'
      responses:
        '201':
          $ref: '#/components/responses/Created'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...